+++ CHANGELOG.md
@@ -7,6 +7,13 @@ availableAt:
 ---
 
 ### unreleased
+#### Noelware.Violet
+- Add **violet::Some**, which acts like **violet::Ok** but for optional values ([`@auguwu`])
+
+#### Noelware.Violet.Experimental
+- Add **TypeId** for safe, no-RTTI-enabled type indexes ([`@auguwu`])
+- Add **Mono** for a replacement for `std::monostate` ([`@auguwu`])
+
 **Git History**: <https://github.com/Noelware/violet/compare/26.05.01...master>
 
 ### 26.05.01